AtiyahBott.jl

AtiyahBott.jl

The function AtiyahBottFormula

Equivariant Classes

Other Functions

Index

AtiyahBott — Module

AtiyahBott is a module containing an implementation of the Atiyah-Bott residue formula in the Julia language. The theory behind the package is described in the paper "Effective computations of the Atiyah-Bott formula" by Giosuè Muratore and Csaba Schneider (https://arxiv.org/pdf/2105.11183.pdf). The degree of the curves of the moduli space must be at most 13. The projective space must be at most of dimension 254. The colorations (useful to speed up the code) are up to projective spaces of dimension 29. The full list is here: https://github.com/mgemath/Colorations/.

The function AtiyahBottFormula

This is the main function of the package.

AtiyahBott.AtiyahBottFormula — Function

AtiyahBottFormula(n, d, m, P, do_check, show_bar, down_col)

Apply the Atiyah-Bott residue formula to the class P, in the moduli space of rational marked stable maps to the projective space of dimension n of degree d with m marks.

Arguments

n::Int64: the dimension of the projective space, it must be between 1 and 254.

d::Int64: the degree of the stable maps, it must be between 1 and 13.

m::Int64: the number of marks.

P: the equivariant class.

do_check::Bool: if true, checks if P is a well defined zero cycle, and stops the computation if this is not true. If false, the computation may have an unexpected behaviour. By default is true.

show_bar::Bool: hide the progress bar if and only if this condition is false. By default is true.

down_col::Bool: check if all colorations needed in the computation are in the folder Data, and download them otherwise. Once downloaded, those files can be used for future computations. By default is true.

The general construction of P is the following:

julia> P = (g,c,w,s,m) ->

After ->, one has to write an expression in the equivariant classes. All such equivariant classes are functions starting with (g,c,w,s) or (g,c,w,s,m). At the end, they can have more arguments. The expression is a polynomial combination of the equivariant classes. We compute the degree of P by

julia> AtiyahBottFormula(n,d,m,P);

Example

julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,5);
julia> AtiyahBottFormula(3,1,0,P);
Warning: the class is not a 0-cycle.
julia> AtiyahBottFormula(4,1,0,P);
Result: 2875//1
julia> AtiyahBottFormula(4,1,0,P,false);             #same as before, but without the preliminary check on `P`
julia> AtiyahBottFormula(4,1,0,P,false,false);       #same as before, but without showing the progress bar
julia> AtiyahBottFormula(4,1,0,P,false,false,false); #same as before, but without checking for the colorations

The function returns an array of the same dimension of P (non-vectorized classes are assumed as 1-dimensional arrays). The Julia notation for accessing to array is name_of_array[i] where i is an index starting from 1.

Example

julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,2)*Hypersurface(g,c,w,s,3);
julia> x = AtiyahBottFormula(3,2,0,P)[1];
Result: 81//1
julia> x
81//1

The class P supports parameters.

julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,3)*(Incidency(g,c,w,s,2)//3)^(d-1);
julia> d = 2;
julia> AtiyahBottFormula(3,d,0,P);
Result: 27//1
julia> d = 3;
julia> AtiyahBottFormula(3,d,0,P);
Result: 84//1

More examples are available in the support of the equivariant classes. It is enough to type ? and then the name of the class. Currently, the supported classes are:

O1_i (Euler class of $ev^*O(1)$, where $ev$ is a specific evaluation map)

O1 (product of all O1_i)

Psi (cycle of $psi$-classes)

Jet (Euler class of the jet bundle $J^p$ of $ev^*O(q)$)

Hypersurface (Euler class of the direct image of $ev^*O(b)$)

Incidency (cycle parameterizing curves meeting a linear subspace)

Contact (cycle parameterizing contact curves)

R1 (first derived functor of direct image of $ev^*O(-k)$)

Equivariant Classes

AtiyahBott.O1_i — Function

Equivariant class of the pull-back of O(1) with respect to the i-th evaluation map.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

m::marks: the marks.

i::Int64: the evaluation map.

Example

julia> P = (g,c,w,s,m) -> O1_i(g,c,w,s,m,1)*O1_i(g,c,w,s,m,2);
julia> AtiyahBottFormula(2,1,2,P);
Result: 1//1
julia> P = (g,c,w,s,m) -> O1_i(g,c,w,s,m,1)^2*Hypersurface(g,c,w,s,2);
julia> AtiyahBottFormula(3,1,1,P);
Result: 4//1

Attention!

The program will stop if i is not between 1 and the number of marks.

AtiyahBott.O1 — Function

Equivariant class of the pull-back of O(1) with respect to the product of all evaluation maps.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

m::marks: the marks.

This function is equivalent to the product of the function O1_i(g,c,w,s,m,i) where i runs from 1 to the number of marks.

Example

julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m);
julia> AtiyahBottFormula(2,3,8,P);
Result: 12//1
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*Hypersurface(g,c,w,s,3);
julia> AtiyahBottFormula(3,2,1,P);
Result: 81//1

In order to remove O1_i(g,c,w,s,m,i) for some i, it is enough to divide by that function. # Example

julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)//O1_i(g,c,w,s,m,1);

Here P is the product of all O1_i(g,c,w,s,m,i) where i runs from 2 to m.

AtiyahBott.Psi — Function

Equivariant class of the cycle of $psi$-classes.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

m::marks: the marks.

a::Vector{Int64}: the vector of the exponents of the psi classes. It is ordered, meaning that the first element is the exponent of $psi_1$, the second is the exponent of $psi_2$, and so on.

Note

The size of a must be at most m. If it is smaller, missing exponents will be considered as zeros. If a is a number, it will be considered as the exponent of $psi_1$.

Attention!

The program will stop if we have one of the following conditions:

the size of a is bigger than m,

a contains a negative number.

Example

julia> P = (g,c,w,s,m) -> O1_i(g,c,w,s,m,1)^5*O1_i(g,c,w,s,m,2)^2*Hypersurface(g,c,w,s,5)*Psi(g,c,w,s,m,[1,0]);
julia> AtiyahBottFormula(6,2,2,P);
Result: 495000//1
julia> P = (g,c,w,s,m) -> O1_i(g,c,w,s,m,1)^8*O1_i(g,c,w,s,m,2)^6*Hypersurface(g,c,w,s,7)*Psi(g,c,w,s,m,2);
julia> AtiyahBottFormula(10,2,2,P);
Result: 71804533752//1
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*Psi(g,c,w,s,m,4);
julia> AtiyahBottFormula(2,2,1,P);
Result: 1//8
julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,2)^4*O1_i(g,c,w,s,m,1)*(O1_i(g,c,w,s,m,1) + Psi(g,c,w,s,m,1))
julia> AtiyahBottFormula(2,2,1,P); #number of plane conics through four points and tangent to a line
Result: 2

Psi is singleton!

Psi cannot be multiplied by itself.

julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*Psi(g,c,w,s,m,1)^4;                  #this is **wrong**
julia> AtiyahBottFormula(2,2,1,P);
Warning: more instances of Psi has been found. Type:
julia> ?Psi
for support.
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*Psi(g,c,w,s,m,3)*Psi(g,c,w,s,m,1);   #this is **wrong**
julia> AtiyahBottFormula(2,2,1,P);
Warning: more instances of Psi has been found. Type:
julia> ?Psi
for support.
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*Psi(g,c,w,s,m,4);
julia> AtiyahBottFormula(2,2,1,P);
Result: 1//8
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)^2*(Psi(g,c,w,s,m,3)*O1(g,c,w,s,m)+Psi(g,c,w,s,m,2)*O1(g,c,w,s,m)^2);
julia> AtiyahBottFormula(2,2,1,P);
Result: 1//8
julia> P = (g,c,w,s,m) -> O1(g,c,w,s,m)*(Psi(g,c,w,s,m,7)*O1(g,c,w,s,m)+Psi(g,c,w,s,m,6)*O1(g,c,w,s,m)^2);
julia> AtiyahBottFormula(3,2,1,P);
Result: -5//16

AtiyahBott.Jet — Function

Equivariant class of the jet bundle J^p of the pull back of O(q) with respect to the first $psi$-class.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

m::marks: the marks.

p::Int64: the exponent of the Jet bundle. In particular, it is a bundle of rank p+1.

q::Int64: the degree of the line bundle that is pulled back.

Note

In order to define this bundle, the number of marks must be at least 1. You cannot multiply this bundle by the class $Psi$.

Example

julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,2)^4*Jet(g,c,w,s,m,1,1);
julia> AtiyahBottFormula(2,2,1,P);
Result: 2//1
julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,2)^4*(Jet(g,c,w,s,m,1,1)+O1(g,c,w,s,m)^2);
julia> AtiyahBottFormula(2,2,1,P);
Result: 3//1
julia> P = (g,c,w,s,m) -> (O1(g,c,w,s,m)^2)//k*Jet(g,c,w,s,m,4*d-2,k);
julia> d=1;k=1;AtiyahBottFormula(3,d,1,P);   #The value of this integral does not depend on k, only on d

AtiyahBott.Hypersurface — Function

Equivariant class of the Euler class of the bundle equal to the direct image under the forgetful map of ev^*O(b). It parameterizes curves contained in a hypersurface of degree b.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

b::Int64: the degrees of the hypersurface. Alternatively, it can be an array of integers, meaning the multiplication of the equivariant class defined by each element of the array.

Example

julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,5);
julia> AtiyahBottFormula(4,1,0,P);
Result: 2875//1
julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,[3,3]);
julia> AtiyahBottFormula(5,2,0,P);
Result: 423549//8
julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,4)*Hypersurface(g,c,w,s,2);
julia> AtiyahBottFormula(5,3,0,P);
Result: 422690816//27
julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,2)^4;
julia> AtiyahBottFormula(7,4,0,P);
Result: 25705160//1

Attention!

The program will stop if b is not positive.

AtiyahBott.Incidency — Function

Equivariant class of the cycle parameterizing curves meeting a linear subspace of codimension r.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

r::Int64: the codimension of the subvariety. Alternatively, it can be an array of integers, meaning the multiplication of the equivariant class defined by each element of the array.

Example

julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,3)^2;
julia> AtiyahBottFormula(3,1,0,P);
Result: 1//1
julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,[2,2,3]);
julia> AtiyahBottFormula(3,1,0,P);
Result: 1//1
julia> P = (g,c,w,s,m) -> Incidency(g,c,w,s,[2,2])*Hypersurface(g,c,w,s,3);
julia> AtiyahBottFormula(3,3,0,P);
Result: 756//1

Attention!

The program will stop if r is not positive.

AtiyahBott.Contact — Function

Equivariant class of the Euler class of the bundle equal to the direct image under the forgetful map of: ev^*O(2) tensor the dualizing sheaf of the forgetful map. It parameterizes contact curves in an odd dimensional projective space.

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

Example

julia> P = (g,c,w,s,m) -> O1_i(g,c,w,s,m,1)^2*O1_i(g,c,w,s,m,2)^3*Contact(g,c,w,s);
julia> AtiyahBottFormula(3,1,2,P);
Result: 1//1

AtiyahBott.R1 — Function

The equivariant class of the first derived functor of the pull-back of O(-k).

Arguments

g::SimpleGraph: the graph.

c::Vector{UInt8}: the coloration.

w::Vector{Int64}: the weights.

s::Rational{BigInt}: the scalars.

k::Int64: a positive integer.

Example

julia> P = (g,c,w,s,m) -> R1(g,c,w,s,1)^2;
julia> AtiyahBottFormula(1,3,0,P);
Result: 1//27

Attention!

The program will stop if k is not positive.

Other Functions

AtiyahBott.dim_M — Function

dim_M(n, d, m)

The dimension of the moduli space of stable rational map to the projective space of dimension n, of degree d with m marks.

Arguments

n::Int64: the dimension of the projective space.

d::Int64: the degree of the stable maps.

m::Int64: the number of marks.

Example

julia> dim_M(2,2,5)
10

AtiyahBott.codim — Function

codim(n, d, m, P)

The codimension of the equivariant class P.

Arguments

n::Int64: the dimension of the projective space.

d::Int64: the degree of the stable maps.

m::Int64: the number of marks.

P: the equivariant class.

Example

julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,5);
julia> codim(4,1,0,P)
6

AtiyahBott.is_zero_cycle — Function

is_zero_cycle(n, d, m, P)

Return true if the equivariant class P is a 0-cycle in the moduli space, false otherwise.

Arguments

n::Int64: the dimension of the projective space.

deg::Int64: the degree of the stable maps.

n_marks::Int64: the number of marks.

P: the equivariant class.

Example

julia> P = (g,c,w,s,m) -> Hypersurface(g,c,w,s,5);
julia> is_zero_cycle(4,1,0,P)
true

AtiyahBott.check_Data — Function

check_Data()

List of all files containing the colorations in the folder Data.

AtiyahBott.fill_Data — Function

fill_Data(n, d)

Download from internet all colorations used for computations in the moduli space of with relative n and d. Return true if there is no need to download any file or if all downloads go well. Otherwise, return false.

AtiyahBott.AtiyahBottFormulaForGraph — Function

AtiyahBottFormulaForGraph(g, pruf_str, aut, n, deg, n_marks, P, s)

Apply the Atiyah-Bott formula to all colorations of a specific graph. It is useful for splitting the computation in multiple parts, to be computed in single threads.

Index

AtiyahBott

AtiyahBott.AtiyahBottFormula

AtiyahBott.AtiyahBottFormulaForGraph

AtiyahBott.Contact

AtiyahBott.Hypersurface

AtiyahBott.Incidency

AtiyahBott.Jet

AtiyahBott.O1

AtiyahBott.O1_i

AtiyahBott.Psi

AtiyahBott.R1

AtiyahBott.check_Data

AtiyahBott.codim

AtiyahBott.dim_M

AtiyahBott.fill_Data

AtiyahBott.is_zero_cycle